Revert "float instead of double"
This reverts commit c06048f7edd6af2043469bdaa6195003448cedb8.

+++ src/models/ppm.cpp
@@ -138,7 +138,7 @@ void PPM::ByteUpdate() {
     node = &tables_[node->lower_table];
     --order;
   }
-  float sum = 0, min_value = 0.0000001;
+  double sum = 0, min_value = 0.0000001;
   for (int i = 0; i < 256; ++i) {
     if (probs_[i] < min_value) probs_[i] = min_value;
     sum += probs_[i];
